Add Votable module, include in Recording & Landmark
