+++ app/models/recording.rb
@@ -1,4 +1,6 @@
 class Recording < ApplicationRecord
+  include Votable
+  
   belongs_to :user
   belongs_to :landmark
   has_many :tour_recordings
@@ -7,9 +9,4 @@ class Recording < ApplicationRecord
 
   validates_presence_of :title, :url
   validates_uniqueness_of :url
-
-  def total_score
-    recording = RecordingDecorator.new(self)
-    recording.total_score
-  end
 end
